rubyx: sepolicy: Re-label `Charging_Adapter`
To use the value of usbpd_verified as the OEM fast charger status path, we can't have it within the  `sysfs_batteryinfo` context.
In current ROMs, system_server attempts to read the value of the file. However, AOSP sepolicy has a neverallow for system_server to read anything labeled with `sysfs_batteryinfo:file`, so simply trying to allow it would fail to build.

As a workaround i created a custom context and allowed it in the same domains where sysfs_batteryinfo was already being accessed.

Signed-off-by: YagizErdemir06 <erdemiryagiz@proton.me>